+++ DataScience/Core/Vector/AccessSlice.cs
